+++ libs/dot-components/src/lib/components/split-button/SplitButton.spec.tsx
@@ -4,9 +4,9 @@ import { render, screen } from '../../testing-utils';
 import { DotSplitButton, SplitButtonProps } from './SplitButton';
 
 const options = [
-  { children: 'option 1', key: '0' },
-  { children: 'option 2 with some longer text', key: '1' },
-  { children: 'option 3', key: '2' },
+  { children: 'option 1', key: 'a' },
+  { children: 'option 2 with some longer text', key: 'b' },
+  { children: 'option 3', key: 'c' },
 ];
 
 const onSelect = jest.fn();
@@ -15,11 +15,13 @@ describe('DotButton', () => {
   it('should have unchanged API', () => {
     const props = {
       ariaLabel: 'splitsville',
+      disablePortal: true,
       onSelect: onSelect,
       options: options,
     };
     const splitButtonProps: SplitButtonProps = {
       ariaLabel: 'splitsville',
+      disablePortal: true,
       onSelect: onSelect,
       options: options,
     };
@@ -48,6 +50,21 @@ describe('DotButton', () => {
     const buttons = screen.getAllByRole('button');
     userEvent.click(buttons[0]);
     expect(onClick).toHaveBeenCalledTimes(1);
+  });
+
+  it('should allow me to select menu item with non-numeric key', () => {
+    render(
+      <DotSplitButton
+        ariaLabel="splitsville"
+        onSelect={onSelect}
+        options={options}
+      />
+    );
+    const buttons = screen.getAllByRole('button');
+    userEvent.click(buttons[1]);
+    const listItems = screen.getAllByRole('menuitem');
+    userEvent.click(listItems[0]);
+    expect(onSelect).toHaveBeenCalledTimes(1);
   });
 
   it('should not allow me to click disabled button', () => {