+++ libs/dot-components/src/lib/components/split-button/SplitButton.styles.ts
@@ -1,5 +1,6 @@
 import { ButtonGroup } from '@material-ui/core';
 import styled from 'styled-components';
+import { DotMenu } from '../menu/Menu';
 
 export const rootClassName = 'dot-split-button-group';
 
@@ -8,5 +9,11 @@ export const StyledSplitButtonGroup = styled(ButtonGroup)`
     .dot-button {
       margin: 0;
     }
+  }
+`;
+
+export const StyledMenu = styled(DotMenu)`
+  &.dot-menu {
+    z-index: 9999;
   }
 `;